refactor PhotoFrame to PhotoWall and PhotoList
